hopefully improved the code
