+++ examples/stop_and_persist.rs
@@ -1,9 +1,8 @@
-use spinoff::Spinners;
-use std::thread::sleep;
-use std::time::Duration;
+use spinoff::{Spinner, Spinners};
+use std::{thread::sleep, time::Duration};
 
 fn main() {
-    let sp = spinoff::new(Spinners::Dots, "Loading...", None);
+    let sp = Spinner::new(Spinners::Dots, "Loading...", None);
     sleep(Duration::from_secs(5));
     sp.stop_and_persist("🍕", "Pizza!");
 }
