--- src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod printer;
-pub mod spinner_data;
-pub mod spinner_enum;
